requirements.txt
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
decorator==5.2.1
Django==5.2.4
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
executing==2.2.0
inflection==0.5.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
phonenumbers==9.0.10
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg-binary==3.2.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.14.1
uritemplate==4.2.0
wcwidth==0.2.13
